+++ docs/changelog.md
@@ -1,3 +1,7 @@
+## Sprint 5
+Ends October 14, 2020
+- add declarations.d.ts for global types.
+
 ## Sprint 4
 Ends September 16, 2020